Move to a seperated folder for Views
